requirements.txt
alabaster==0.7.12
anyascii==0.1.7
appdirs==1.4.4
asgiref==3.3.1
astor==0.8.1
attrs==20.3.0
autoflake==1.4
autopep8==1.5.5
Babel==2.9.0
bandit==1.7.0
beautifulsoup4==4.8.2
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.4
darglint==1.7.0
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-toolbelt==0.0.1
django-treebeard==4.4
djangorestframework==3.12.2
doc8==0.8.1
docutils==0.16
draftjs-exporter==2.1.7
eradicate==1.0
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
gitdb==4.0.5
GitPython==3.1.13
gunicorn==20.0.4
html5lib==1.1
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==4.3.21
jdcal==1.4.1
Jinja2==2.11.3
l18n==2020.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
openpyxl==3.0.6
packaging==20.9
pathspec==0.8.1
pbr==5.5.1
pep8-naming==0.9.1
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.8.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
python-dev-tools==2020.9.10
pytz==2021.1
pyupgrade==2.10.0
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
restructuredtext-lint==1.3.2
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
soupsieve==2.2
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
static3==0.7.0
stevedore==3.3.0
tablib==3.0.0
testfixtures==6.17.1
tokenize-rt==4.1.0
toml==0.10.2
tox==3.22.0
tox-travis==0.12
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
wagtail==2.12.2
webencodings==0.5.1
wemake-python-styleguide==0.14.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
